backend/requirements.txt
flask
flask-cors
flask-jwt-extended
flask-bcrypt
python-dotenv
supabase
requests